requirements.txt
streamlit
transformers
torch
python-pptx
PyMuPDF
pytesseract
Pillow
opencv-python
PyPDF2
